requirements.txt
PyYAML
rich
numpy
